barcode processing changed
when barcodes don't need to be corrected for
sequencing errors, they are assumed to be
encoded in the sequencing read names.
now the barcodes are extracted from the read name
and matched up with a set of desired barcode names
from the barcodesheet.tsv.
min_mapq and max_mismatch are now associated
to barcodes to enhance clarity.