added logging, fixed issues for counting
a log file is produced, summarizing the tag counts.
before, when both mates of a paired-end fragment started
on the same position, they would be skipped.
now, in order to just count each pair once, the first
read is considered only and the read start and template length
adjusted accordingly.